requirements.txt
Django
Pillow
mysqlclient
django-crispy-forms
python-dateutil
django-registration-redux
